Extract existing gulpfile tasks into discrete tasks. Use run-sequence to avoid race condition. Add clean task and run that first to avoid stale compiled assets being used. Fix invalid HTML ('clas' instead of 'class' in index.html). Next steps: use proper packages for bootstrap and modernizr to eliminate noise they add to ESLint output, prettify ESLint output.
