+++ .gitignore
@@ -2,3 +2,4 @@
 *.swo
 .DS_Store
 node_modules
+.idea
