+++ index.html
@@ -13,7 +13,7 @@
   <script src="js/modernizr-custom.js"></script>
 
   <link href='https://fonts.googleapis.com/css?family=Roboto+Mono|Roboto' rel='stylesheet' type='text/css'>
-  <link href="css/build.min.css" rel="stylesheet">
+  <link href="dist/dist.min.css" rel="stylesheet">
 
   <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
   <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
@@ -449,7 +449,7 @@
             <h4 class="modal-title" id="importModalLabel">Import</h4>
           </div>
           <div class="modal-body clearfix text-center">
-            <div clas="row">
+            <div class="row">
               <div class="col-xs-6 col-sm-3">
                 <span class="gem gem0"></span> = 0 / r<br/>
                 <span class="gem gem1"></span> = 1 / b<br/>
@@ -484,7 +484,7 @@
             <h4 class="modal-title" id="changeOrbsModalLabel">Change Orbs</h4>
           </div>
           <div class="modal-body clearfix text-center">
-            <div clas="row">
+            <div class="row">
               <div class="col-xs-6">
                 <span class="gem gem-lg gem0"></span> &rarr; <span id="orginal-gem0" class="gem gem-lg gem0 gem-target"></span><br />
                 <span class="gem gem-lg gem1"></span> &rarr; <span id="orginal-gem1" class="gem gem-lg gem1 gem-target"></span><br />
@@ -523,6 +523,6 @@
 
   <script src="https://code.jquery.com/jquery-1.12.2.min.js" integrity="sha256-lZFHibXzMHo3GGeehn1hudTAP3Sc0uKXBXAzHX1sjtk=" crossorigin="anonymous"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/camanjs/4.1.2/caman.full.min.js"></script>
-  <script src="js/build.min.js"></script>
+  <script src="dist/dist.min.js"></script>
 </body>
 </html>
