+++ tasks/default.js
@@ -1,0 +1,5 @@
+const gulp = require('gulp');
+const runSequence = require('run-sequence').use(gulp);
+
+gulp.task('default', function() { runSequence('clean', 'sass', 'javascript', 'server', 'watch') } );
+
